🐫 [ui] rename files to proud camel
y'all are crazy but, at least it's standardized? 😖
