+++ emoji-ui/src/components/EmojiItem.css
