+++ emoji-ui/src/components/EmojiItem.js
