+++ emoji-ui/src/store/ActionTypes.js
