+++ emoji-ui/src/store/reducers/SourceEmoji.js
@@ -1,4 +1,4 @@
-import ActionTypes from '../action_types';
+import ActionTypes from '../ActionTypes';
 
 const DEFAULT_STATE = {
   filterStr: '',
