+++ emoji-ui/src/store/selectors/SourceEmoji.js
